+++ services/login/include/service/login/model/challenge_session.hpp
@@ -1,14 +1,12 @@
 #ifndef DARK_TRIAD_LOGIN_SERVICE_CHALLENGE_SESSION_HPP
 #define DARK_TRIAD_LOGIN_SERVICE_CHALLENGE_SESSION_HPP
 
-#include <utility>
-
 class ChallengeSession {
 public:
     std::string id;
-    int challenge;
+    int64_t challenge;
 
-    ChallengeSession(std::string id, int challenge) : id(std::move(id)), challenge(challenge)
+    ChallengeSession(std::string id, int64_t challenge) : id(std::move(id)), challenge(challenge)
     {}
 };
